Remove Models folder from directory and move User entity to domain layer
