+++ app/Domain/BusinessEntity.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Domain;
+
+interface BusinessEntity
+{
+}
